chore: move utility func updateConfigValue to store/client/actions
