+++ src/store/client/pluginSideEffects.js
@@ -1,24 +1,6 @@
 import Clef from '../requests/clefService'
-import { setConfig } from './actions'
+import { updateConfigValue } from './actions'
 import { networkToChainId } from '../../lib/utils'
-
-export const updateConfigValue = (clientName, key, value) => {
-  return (dispatch, getState) => {
-    const client = getState().client[clientName]
-    if (!client) {
-      return
-    }
-    const { config } = client
-    if (config[key] === value) {
-      return
-    }
-    const newConfig = {
-      ...config,
-      [key]: value
-    }
-    dispatch(setConfig(clientName, newConfig))
-  }
-}
 
 export const startClientSideEffects = clientName => {
   return dispatch => {
